requirements.txt
certifi==2019.11.28
chardet==3.0.4
DateTime==4.3
idna==2.9
nest-asyncio==1.3.0
pycodestyle==2.5.0
pyflakes==2.1.1
pyodbc==4.0.30
python-dateutil==2.8.1
pytz==2019.3
requests==2.23.0
selenium==3.141.0
six==1.14.0
splinter==0.13.0
urllib3==1.25.8
websockets==8.1
zope.interface==4.7.2
